Add --no-verify in commit for submodule travis file
When we run the task to add pull requests in a submodule
(submodule.merges), it adds a commit to exclude travis builds.
If pre-commit is installed in the repository, it will be run during
the invoke task. Use --no-verify to ignore it.

---------

This commit has been extracted from https://github.com/camptocamp/odoo-template