🐛 Let subprocess.run access venv binaries
When installing this project in a venv (as it should be), the dependencies
are not directly exposed in the PATH.

So, CLI dependencies like `bumpversion`, `invoke`, and others are not found
when running `subprocess.run` in the code.

This commit fixes this by adding the venv binaries to the PATH. Since they're
added at the beginning of the PATH, they will be found before any other
system-wide binaries, which ensures the correct version is used.