add support for Odoo 17 packaging
The OCA uses wool to package addons on odoo >= 17. This has the side effect of
changing the string used to install an addon from github (the path to use is slightly
different.

We add some command line options to force the usage of wool, and try to autodetect (by
checking the project version).