[ADD] otools-ba: tools for business analysts
These tools are not necessarily directly useful for projects, but they
can be used to abstract things out, and provide direct shortcuts for BAs
who may not be super well versed in the internals of git, docker, etc.

First command provided: ``otools-ba run <odoo-version>``

This will download a "new" docker image for that version and start it.
Some options are proviced to set a port, preserve a possibly preexisting
database, and pull an updated version of the image.