[IMP] change the way we detect the odoo commit hash
The previous version was using the projects docker composition. This
would fail if the docker composition has not been built, and would not
get the correct versions if the project had been built a long time ago
with an out of date base image. Also during project conversion, this
would lead to issues because the docker image may not be buildable,
because of the presence of patches which cannot be applied, and fixing
these requires to get the hashes to be able to add the pending merges on
odoo and enterprise.

This PR changes the way to run by using a standard docker run, with the
pulling of the image forced.