[FIX] browser being launched more than once
If a worker restart, possibly because of memory consumption, the the
line which triggers launching the browser will be redisplayed in the
logs, and the tool will restart the browser.

We fix this by stopping the search for the "Registry loaded" string
after we find it for the first time.

fixes: https://github.com/camptocamp/odoo-project-tools/issues/90