[FIX] otools-migrate-db: avoid extra DB removal and restore with `--restart-c2c` flag set
When the parameter `--restart-c2c` is set, we wish to remove current
working DB and restore it in 'core' step (with the DB coming from the last
step, so `odoo_migrated` here).

Then for next steps, we don't want anymore to drop & restore the DB
because the `--restart-c2c` flag is set. Working on current DB is OK.
The only case where DB has to be dropped is then the flag
`--restart-c2c-{step}` is set.