[IMP] checkout-local-odoo: compatible with odoosh-template
We don't use a VERSION file in odoosh-template, we don't have to.

In fact the version here was used to determine the branch to clone, but as soon
as we clone we are checking out a specific commit anyway, so it's not relevant.

Instead, we will clone with --no-checkout, and fetch only the target ref.